Move helper modules to client/lib/
