+++ client/components/basic/userStatus/UserStatus.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Box } from '@rocket.chat/fuselage';
 
-import statusColors from '../../../helpers/statusColors';
+import statusColors from '../../../lib/statusColors';
 
 const UserStatus = React.memo(({ status, ...props }) => <Box size='x12' borderRadius='full' backgroundColor={statusColors[status]} {...props}/>);
