+++ client/lib/capitalize.ts
