+++ client/lib/createSidebarItems.ts
