+++ client/lib/getUserEmailAddress.js
