+++ ee/app/engagement-dashboard/client/components/UsersTab/ActiveUsersSection.js
@@ -9,7 +9,7 @@ import CounterSet from '../../../../../../client/components/data/CounterSet';
 import { LegendSymbol } from '../data/LegendSymbol';
 import { Section } from '../Section';
 import { ActionButton } from '../../../../../../client/components/basic/Buttons/ActionButton';
-import { downloadCsvAs } from '../../../../../../client/helpers/download';
+import { downloadCsvAs } from '../../../../../../client/lib/download';
 
 export function ActiveUsersSection() {
 	const t = useTranslation();
