+++ ee/app/engagement-dashboard/client/components/UsersTab/NewUsersSection.js
@@ -8,7 +8,7 @@ import { useEndpointData } from '../../../../../../client/hooks/useEndpointData'
 import CounterSet from '../../../../../../client/components/data/CounterSet';
 import { Section } from '../Section';
 import { ActionButton } from '../../../../../../client/components/basic/Buttons/ActionButton';
-import { downloadCsvAs } from '../../../../../../client/helpers/download';
+import { downloadCsvAs } from '../../../../../../client/lib/download';
 
 export function NewUsersSection() {
 	const t = useTranslation();
